Removed authentication logic from user model
